# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Araldo API
  version: v1
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  description: |
    Schedule and publish posts to every platform from your product.

    **Authentication.** Send an API key as a bearer token:
    `Authorization: Bearer ald_test_…`. Test keys (`ald_test_`) only ever
    reach sandbox channels; live keys (`ald_live_`) reach real accounts.

    **Idempotency.** Every `POST` accepts an `Idempotency-Key` header. A
    retry with the same key and body returns the first response (with
    `Idempotent-Replayed: true`) for 24 hours.

    **Errors** are RFC 9457 problem details with a stable `code`, the
    `param` at fault, and every validation problem in `errors`.

    **Lists** page by cursor: `limit`, `starting_after`, `ending_before`,
    and `has_more` in the response.

    **Events and webhooks.** Every change of interest becomes an event,
    listed at `/v1/events` and delivered to your webhook endpoints, signed
    with `Araldo-Signature: t=<unix>,v1=<hex HMAC-SHA256 of "t.body">`.
servers:
  - url: /
security:
  - apiKey: []
...
paths:
...
  /v1/platforms:
    get:
      summary: List platforms and their rules
      description: Length limits, how length is counted, threads and media rules for every platform. In live mode, also the fields needed to connect a channel.
      tags: [Platforms]
      responses:
        "200":
          description: Platforms.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ListBase"
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: "#/components/schemas/Platform" } }
        default: { $ref: "#/components/responses/Error" }
  /v1/brands:
    get:
      summary: List brands
      tags: [Brands]
      responses:
        "200":
          description: Brands.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ListBase"
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: "#/components/schemas/Brand" } }
        default: { $ref: "#/components/responses/Error" }
    post:
      summary: Create a brand
      description: New brands get weekday publishing slots at 09:00 and 13:00 in their time zone.
      tags: [Brands]
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BrandInput" }
      responses:
        "201":
          description: The brand.
          content: { application/json: { schema: { $ref: "#/components/schemas/Brand" } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/brands/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: A brand ID (`brand_…`) or slug.
        schema: { type: string }
    get:
      summary: Retrieve a brand
      tags: [Brands]
      responses:
        "200":
          description: The brand.
          content: { application/json: { schema: { $ref: "#/components/schemas/Brand" } } }
        default: { $ref: "#/components/responses/Error" }
    post:
      summary: Update a brand
      tags: [Brands]
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BrandInput" }
      responses:
        "200":
          description: The brand.
          content: { application/json: { schema: { $ref: "#/components/schemas/Brand" } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/channels:
    get:
      summary: List channels
      description: Channels in the key's mode.
      tags: [Channels]
      parameters:
        - { name: brand, in: query, schema: { type: string }, description: Brand ID or slug. }
      responses:
        "200":
          description: Channels.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ListBase"
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: "#/components/schemas/Channel" } }
        default: { $ref: "#/components/responses/Error" }
    post:
      summary: Connect a channel
      description: |
        Araldo checks the credentials with the platform before saving them
        (encrypted). With a test key, only `provider: sandbox` works; give
        the platform it imitates in `emulates`.
      tags: [Channels]
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [brand, provider]
              properties:
                brand: { type: string }
                provider: { type: string, example: sandbox }
                emulates: { type: string, example: bluesky, description: For sandbox channels. }
                fields:
                  type: object
                  additionalProperties: { type: string }
                  description: The platform's connect fields (see /v1/platforms).
      responses:
        "201":
          description: The channel.
          content: { application/json: { schema: { $ref: "#/components/schemas/Channel" } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/channels/{id}:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    get:
      summary: Retrieve a channel
      tags: [Channels]
      responses:
        "200":
          description: The channel.
          content: { application/json: { schema: { $ref: "#/components/schemas/Channel" } } }
        default: { $ref: "#/components/responses/Error" }
    delete:
      summary: Disconnect a channel
      tags: [Channels]
      responses:
        "200":
          description: Deleted.
          content: { application/json: { schema: { $ref: "#/components/schemas/Deleted" } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/templates:
    get:
      summary: List templates
      tags: [Templates]
      parameters:
        - { name: brand, in: query, schema: { type: string } }
      responses:
        "200":
          description: Templates (without their bodies).
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ListBase"
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: "#/components/schemas/Template" } }
        default: { $ref: "#/components/responses/Error" }
    post:
      summary: Create a template
      tags: [Templates]
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/TemplateSource"
                - type: object
                  required: [brand, key]
                  properties:
                    brand: { type: string }
                    key: { type: string, example: featured-build }
                    name: { type: string }
                    approval: { $ref: "#/components/schemas/TemplateApproval" }
      responses:
        "201":
          description: The template with version 1.
          content: { application/json: { schema: { $ref: "#/components/schemas/Template" } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/templates/preview:
    post:
      summary: Preview a template
      description: Render a saved template (`brand` + `template`) or an unsaved one (the source fields) for each platform. Nothing is stored.
      tags: [Templates]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/TemplateSource"
                - type: object
                  properties:
                    brand: { type: string }
                    template: { type: string, description: "key, key@version or template ID" }
                    data: { type: object, additionalProperties: true }
                    providers: { type: array, items: { type: string } }
      responses:
        "200":
          description: One rendition per platform.
          content: { application/json: { schema: { $ref: "#/components/schemas/Preview" } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/templates/{id}:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    get:
      summary: Retrieve a template
      tags: [Templates]
      parameters:
        - { name: version, in: query, schema: { type: integer, minimum: 1 }, description: Default the latest. }
      responses:
        "200":
          description: The template and the version asked for.
          content: { application/json: { schema: { $ref: "#/components/schemas/Template" } } }
        default: { $ref: "#/components/responses/Error" }
    post:
      summary: Update a template's settings
      description: |
        Rename a template or change its approval rule, without making a new
        version. Only people who can approve posts may change `approval`, so
        API keys cannot.
      tags: [Templates]
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                approval: { $ref: "#/components/schemas/TemplateApproval" }
      responses:
        "200":
          description: The template.
          content: { application/json: { schema: { $ref: "#/components/schemas/Template" } } }
        default: { $ref: "#/components/responses/Error" }
    delete:
      summary: Delete a template
      description: Posts made from it keep their text.
      tags: [Templates]
      responses:
        "200":
          description: Deleted.
          content: { application/json: { schema: { $ref: "#/components/schemas/Deleted" } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/templates/{id}/versions:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    post:
      summary: Save a new version
      description: Versions are immutable. Queued posts keep the text they were rendered with.
      tags: [Templates]
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/TemplateSource"
                - type: object
                  properties:
                    name: { type: string }
      responses:
        "201":
          description: The template with its new version.
          content: { application/json: { schema: { $ref: "#/components/schemas/Template" } } }
        default: { $ref: "#/components/responses/Error" }
...
  /v1/posts/preview:
    post:
      summary: Preview a post
      description: Renders a post for each channel without saving it. Rule violations are listed per rendition (not as an error), so an agent can read them and fix its text.
      tags: [Posts]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PostInput" }
      responses:
        "200":
          description: One rendition per channel.
          content: { application/json: { schema: { $ref: "#/components/schemas/Preview" } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/posts/{id}:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    get:
      summary: Retrieve a post
      tags: [Posts]
      responses:
        "200":
          description: The post and its targets.
          content: { application/json: { schema: { $ref: "#/components/schemas/Post" } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/posts/{id}/cancel:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    post:
      summary: Cancel a post
      description: Cancels every target not yet published or publishing.
      tags: [Posts]
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      responses:
        "200":
          description: The post.
          content: { application/json: { schema: { $ref: "#/components/schemas/Post" } } }
        default: { $ref: "#/components/responses/Error" }
  /v1/post_targets/{id}/retry:
    parameters: [{ $ref: "#/components/parameters/ID" }]
    post:
      summary: Retry a target
      description: Queues a failed or `needs_attention` target again. For `needs_attention`, you assert the post did not go out.
      tags: [Posts]
      parameters: [{ $ref: "#/components/parameters/IdempotencyKey" }]
      responses:
        "200":
          description: The post.
          content: { application/json: { schema: { $ref: "#/components/schemas/Post" } } }
        default: { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: "An API key: ald_test_… or ald_live_…"
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Error:
      description: A problem.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    ListBase:
      type: object
      required: [object, data, has_more, url]
      properties:
        object: { type: string, enum: [list] }
        data: { type: array, items: {} }
        has_more: { type: boolean }
        url: { type: string }
...